+++ tests/unittest/analyses/notch/test_notch_config.py
@@ -12,6 +12,14 @@ from perseo_quality.elevation_notch_analysis.config import ElevationNotchConfig
 
 from sct.analyses.elevation_notch.config import SCTElevationNotchAnalysisConfig
 
+general_config_toml = """
+
+[general]
+save_log = true
+save_config_copy = true
+
+"""
+
 elevation_notch_analysis_toml = """
 
 [elevation_notch_analysis]
@@ -84,6 +92,15 @@ class NOTCHConfigurationTest(unittest.TestCase):
             new_config = SCTElevationNotchAnalysisConfig.from_toml(path_to_new_file)
 
         assert new_config == config
+
+    def test_empty_config(self) -> None:
+        """Test empty configuration"""
+        with self.assertRaises(ValidationError):
+            with TemporaryDirectory() as temp_dir:
+                path_to_file = Path(temp_dir).joinpath("test").with_suffix(".toml")
+                path_to_file.write_text(general_config_toml)
+
+                SCTElevationNotchAnalysisConfig.from_toml(path_to_file)
 
 
 if __name__ == "__main__":