+++ tests/unittest/analyses/pta/test_pta_config.py
@@ -15,6 +15,15 @@ from perseo_quality.point_targets_analysis.config import IRFParameters, PointTar
 
 from sct.analyses.point_target.config import SCTPointTargetAnalysisConfig
 
+general_config_toml = """
+
+[general]
+save_log = true
+save_config_copy = true
+
+"""
+
+
 point_target_analysis_toml = """
 
 [point_target_analysis]
@@ -167,6 +176,15 @@ class PTAConfigurationTest(unittest.TestCase):
             new_config = SCTPointTargetAnalysisConfig.from_toml(path_to_new_file)
 
         assert new_config == config
+
+    def test_empty_config(self) -> None:
+        """Test empty configuration"""
+        with self.assertRaises(ValidationError):
+            with TemporaryDirectory() as temp_dir:
+                path_to_file = Path(temp_dir).joinpath("test").with_suffix(".toml")
+                path_to_file.write_text(general_config_toml)
+
+                SCTPointTargetAnalysisConfig.from_toml(path_to_file)
 
 
 if __name__ == "__main__":